+++ tests/charging-station/ocpp/2.0/OCPP20IncomingRequestService-GroupIdStop.test.ts
@@ -11,7 +11,7 @@ import type { OCPP20RequestStartTransactionRequest } from '../../../../src/types
 
 import { createTestableIncomingRequestService } from '../../../../src/charging-station/ocpp/2.0/__testable__/index.js'
 import { OCPP20IncomingRequestService } from '../../../../src/charging-station/ocpp/2.0/OCPP20IncomingRequestService.js'
-import { OCPPAuthServiceFactory } from '../../../../src/charging-station/ocpp/auth/services/OCPPAuthServiceFactory.js'
+import { OCPPAuthServiceFactory } from '../../../../src/charging-station/ocpp/auth/index.js'
 import {
   OCPP20IdTokenEnumType,
   RequestStartStopStatusEnumType,
